#386: move used function of rank_object_enumerator to lbsStatistsics module and deprecate rank_object_enumerator
